requirement.txt
Flask==2.3.3
gunicorn==21.2.0
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.24.3
gdown==4.7.1
